requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
Deprecated
dj-config-url
Django
django-cors-headers
djangorestframework
gunicorn
idna
jwcrypto
oauthlib
packaging
psycopg2-binary
pycparser
pytz
requests
sqlparse
typing_extensions
urllib3
wrapt
